Suspicious-Activity-Detection-Updated/requirements.txt
pillow>=8.0.0
opencv-python>=4.5.0
numpy>=1.21.0
scikit-learn>=1.0.0